Fix some typos and bugs.
In a chef resource, the cookbook_name will automatically refer to the
_caller_ cookbook, not the actual current cookbook. So if you try to use
cookbook_file or template expecting it to source from the cookbook that
the resource is in, that won't work. You have to explicitly name the
provider cookbook.

Evidently this is a feature. https://github.com/chef/chef/issues/3681