Add a helper script to skip the ASG abandon hook.
When troubleshooting issues with instance provisioning, it is often
useful to keep an instance around for inspection. We have a flag file
that we can drop in /etc/login.gov/info/ that instructs provision.sh not
to complete the ASG lifecycle hook with ABANDON in the event that
provisioning fails.

But it's hard to remember this filename when you're racing the clock to
save the instance.

Add a new script, `id-asg-skip-abandon-hook`, that knows the name of
this file and will create it for you.